Add files via upload
This is apython program for Text Mining. BAsically ,  setof input toens is fed into the program in the form of JSON file format. Those input tokens are then preprocessed in the form that are understandable to a machine learning model. It is basically an initial phase to natural language processing. The set of sttements quoted on social media  , may have specific dates , times , acronyms , abbrevations , contractions(doesn't , ain't) , symbols and fractions(1/7,2/8) , that need to be preprocessed into a set of words that are understanble and more realistic to humans. For instance , A sentence quotes , "I was born in the year 2000" needs to be preprocesssed into I was born in the year two thousand. 

Every conversion from number to words , roman to numbers , contractions to proper words has been shown step by step with proper mapping . The words which were unaffected by the preprocessing were returned as self while the affected words were reflected with changes in the output token file.